html/browsers/browsing-the-web/navigating-across-documents/initial-empty-document/iframe-src-aboutblank-wait-for-load.html

<!DOCTYPE html>
<meta charset="UTF-8">
<title>Navigations after iframe with src='about:blank' finished loading</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>
<body></body>
<script>
/*
  When an iframe is created it will contain the initial empty document. However,
  when the iframe's src is set to "about:blank", it will trigger a navigation to
  a non-initial about:blank document, which will replace the initial empty
  document.
  These tests verify the behavior of navigations that happen after the
  non-initial about:blank document finished loading. These navigations should
  not be auto-converted to do replacement like they would if they start on the
  initial empty document.
*/
"use strict";
const url1 = "/common/blank.html?1";

promise_test(async t => {
  const startingHistoryLength = history.length;
  // Create an iframe with src set to about:blank, and wait for it to finish
  // loading. This would trigger and commit a navigation to a non-initial
  // about:blank document.
  const iframe = await insertIframeWithAboutBlankSrcWaitForLoad(t);
  assert_equals(history.length, startingHistoryLength,
    "Inserting iframe with src='about:blank' must not change history.length");

  // At this point the iframe has loaded a new about:blank document that is not
  // the initial empty document.
  // Now, trigger a navigation through the iframe's src attribute. This should
  // behave like a normal navigation and won't be converted to do a replacement.
  iframe.src = url1;
  await waitForLoad(t, iframe, url1);
  assert_equals(history.length, startingHistoryLength + 1,
    "history.length increases after normal navigation from non-initial empty document");
}, "Navigating to a different document with src");

promise_test(async t => {
  const startingHistoryLength = history.length;
  // Create an iframe with src set to about:blank, and wait for it to finish
  // loading. This would trigger and commit a navigation to a non-initial
  // about:blank document.
  const iframe = await insertIframeWithAboutBlankSrcWaitForLoad(t);
  assert_equals(history.length, startingHistoryLength,
    "Inserting iframe with src='about:blank' must not change history.length");

  // At this point the iframe has loaded a new about:blank document that is not
  // the initial empty document.
  // Now, trigger a navigation through location.href. This should behave like a
  // normal navigation and won't be converted to do a replacement.
  iframe.contentWindow.location.href = url1;
  await waitForLoad(t, iframe, url1);
  assert_equals(history.length, startingHistoryLength + 1,
    "history.length increases after normal navigation from non-initial empty document");
}, "Navigating to a different document with location.href");

promise_test(async t => {
  const startingHistoryLength = history.length;
  // Create an iframe with src set to about:blank, and wait for it to finish
  // loading. This would trigger and commit a navigation to a non-initial
  // about:blank document.
  const iframe = await insertIframeWithAboutBlankSrcWaitForLoad(t);
  assert_equals(history.length, startingHistoryLength,
    "Inserting iframe with src='about:blank' must not change history.length");

  // At this point the iframe has loaded a new about:blank document that is not
  // the initial empty document.
  // Now, trigger a navigation through location.assign. This should behave like
  // a normal navigation and won't be converted to do a replacement.
  iframe.contentWindow.location.assign(url1);
  await waitForLoad(t, iframe, url1);
  assert_equals(history.length, startingHistoryLength + 1,
    "history.length increases after normal navigation from non-initial empty document");
}, "Navigating to a different document with location.assign");

promise_test(async t => {
  const startingHistoryLength = history.length;
  // Create an iframe with src set to about:blank, and wait for it to finish
  // loading. This would trigger and commit a navigation to a non-initial
  // about:blank document.
  const iframe = await insertIframeWithAboutBlankSrcWaitForLoad(t);
  assert_equals(history.length, startingHistoryLength,
    "Inserting iframe with src='about:blank' must not change history.length");

  // At this point the iframe has loaded a new about:blank document that is not
  // the initial empty document.
  // Now, trigger a navigation through window.open. This should behave like a
  // normal navigation and won't be converted to do a replacement.
  iframe.contentWindow.open(url1, "_self");
  await waitForLoad(t, iframe, url1);
  assert_equals(history.length, startingHistoryLength + 1,
    "history.length increases after normal navigation from non-initial empty document");
}, "Navigating to a different document with window.open");

promise_test(async t => {
  const startingHistoryLength = history.length;
  // Create an iframe with src set to about:blank, and wait for it to finish
  // loading. This would trigger and commit a navigation to a non-initial
  // about:blank document.
  const iframe = await insertIframeWithAboutBlankSrcWaitForLoad(t);
  assert_equals(history.length, startingHistoryLength,
    "Inserting iframe with src='about:blank' must not change history.length");

  // At this point the iframe has loaded a new about:blank document that is not
  // the initial empty document.
  // Now, trigger a navigation through a link click. This should behave like a
  // normal navigation and won't be converted to do a replacement.
  const a = iframe.contentDocument.createElement("a");
  a.href = url1;
  iframe.contentDocument.body.appendChild(a);
  a.click();
  await waitForLoad(t, iframe, url1);
  assert_equals(history.length, startingHistoryLength + 1,
    "history.length increases after normal navigation from non-initial empty document");
}, "Navigating to a different document with link click");

promise_test(async t => {
  const startingHistoryLength = history.length;
  // Create an iframe with src set to about:blank which will commit an about:blank document that is not the initial empty document, and wait for it to load.
  const iframe = await insertIframeWithAboutBlankSrcWaitForLoad(t);
  assert_equals(history.length, startingHistoryLength,
    "Inserting iframe with src='about:blank' must not change history.length");

  // At this point the iframe has loaded a new about:blank document that is not
  // the initial empty document.
  // Now, trigger a navigation through a form submission. This should behave
  // like a normal navigation and won't be converted to do a replacement.
  const form = iframe.contentDocument.createElement("form");
  form.action = "/common/blank.html";
  iframe.contentDocument.body.appendChild(form);
  const input = iframe.contentDocument.createElement("input");
  input.type = "hidden";
  input.name = "1";
  form.append(input);
  form.submit();
  await waitForLoad(t, iframe, url1 + "=");
  assert_equals(history.length, startingHistoryLength + 1,
    "history.length increases after normal navigation from non-initial empty document");
}, "Navigating to a different document with form submission");
</script>
